backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
openai==0.28.0
email-validator==2.1.0
python-multipart==0.0.6
asyncpg==0.29.0
alembic==1.12.1
pytest==7.4.3
